move PublicKey and KeyPair types to module - breaking change for KDF imports
