+++ src/blake2b_internal.rs
@@ -1,7 +1,6 @@
 use crate::specifier::Specifier;
 use crate::spend_policy::UnlockKey;
-use crate::types::Hash256;
-use crate::PublicKey;
+use crate::types::{Hash256, PublicKey};
 use blake2b_simd::Params;
 use std::default::Default;
